Extract shared fake DB-session test doubles to tests/_fakes.py (#299)
