+++ tests/test_admin_like_escape.py
@@ -18,6 +18,7 @@ import pytest
 from sqlalchemy.dialects import postgresql
 
 from server.api.admin import _like_escape
+from tests._fakes import install_fake_session_factory
 
 
 def test_percent_is_escaped():
@@ -47,52 +48,6 @@ def test_empty_string_is_unchanged():
     assert _like_escape("") == ""
 
 
-class _FakeScalarRows:
-    def all(self):
-        return []
-
-
-class _FakeResult:
-    def all(self):
-        return []
-
-    def scalars(self):
-        return _FakeScalarRows()
-
-
-class _FakeSession:
-    def __init__(self):
-        self.statements = []
-
-    async def scalar(self, statement):
-        self.statements.append(statement)
-        return 0
-
-    async def execute(self, statement):
-        self.statements.append(statement)
-        return _FakeResult()
-
-
-class _FakeSessionContext:
-    def __init__(self, session):
-        self.session = session
-
-    async def __aenter__(self):
-        return self.session
-
-    async def __aexit__(self, exc_type, exc, tb):
-        return False
-
-
-def _install_fake_session_factory(monkeypatch):
-    session = _FakeSession()
-    monkeypatch.setattr(
-        "server.db.engine.get_session_factory",
-        lambda: lambda: _FakeSessionContext(session),
-    )
-    return session
-
-
 def _compiled(statement):
     return statement.compile(dialect=postgresql.dialect())
 
@@ -105,7 +60,7 @@ def _param_values(compiled):
 async def test_admin_subjects_search_and_internal_prefixes_use_like_escape(monkeypatch):
     from server.api.admin import list_subjects_admin
 
-    session = _install_fake_session_factory(monkeypatch)
+    session = install_fake_session_factory(monkeypatch)
 
     await list_subjects_admin(search="user_123", limit=1, offset=0)
 
@@ -122,7 +77,7 @@ async def test_admin_subjects_search_and_internal_prefixes_use_like_escape(monke
 async def test_admin_memory_search_uses_like_escape(monkeypatch):
     from server.api.admin import list_subject_memories
 
-    session = _install_fake_session_factory(monkeypatch)
+    session = install_fake_session_factory(monkeypatch)
 
     await list_subject_memories("subject-1", search="100%_match", limit=1, offset=0)
 
@@ -136,7 +91,7 @@ async def test_admin_memory_search_uses_like_escape(monkeypatch):
 async def test_admin_episode_search_uses_like_escape(monkeypatch):
     from server.api.admin import list_subject_episodes
 
-    session = _install_fake_session_factory(monkeypatch)
+    session = install_fake_session_factory(monkeypatch)
 
     await list_subject_episodes("subject-1", search="sess_100%", limit=1, offset=0)
